be/requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
networkx==3.1
numpy==1.26.4
openvino==2025.2.0
openvino-telemetry==2025.2.0
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
